Remove "sp" from subcrate names
