+++ cid/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "cid"
+version = "0.1.0"
+edition = "2018"
+
+[package.metadata.release]
+pre-release-commit-message = "Release {{version}} 🎉🎉"
+no-dev-version = true
+
+[features]
+default = ["multibase"]
+std = ["multibase", "multihash/std", "unsigned-varint/std"]
+arb = ["quickcheck", "rand", "multihash/arb"]
+scale-codec = ["parity-scale-codec", "multihash/scale-codec"]
+serde-codec = ["serde", "multihash/serde-codec"]
+
+[dependencies]
+unsigned-varint = { version = "0.7.0", default-features = false }
+
+multibase = { version = "0.9.1", default-features = false, optional = true }
+parity-scale-codec = { version = "2.1.1", default-features = false, features = ["derive"], optional = true }
+quickcheck = { version = "0.9.2", optional = true }
+serde = { version = "1.0.116", optional = true }
+rand = { version = "0.7.3", optional = true }
+sp-std = { version = "3", default-features = false }
+bytecursor = { path = "../bytecursor" }
+multihash = { path = "../multihash" }
+
+[dev-dependencies]
+serde_json = "1.0.59"
