+++ multihash/Cargo.toml
@@ -1,0 +1,52 @@
+[package]
+name = "multihash"
+version = "0.1.0"
+edition = "2018"
+
+[features]
+default = ["derive", "multihash-impl", "secure-hashes"]
+std = ["unsigned-varint/std", "multihash-derive/std"]
+multihash-impl = ["derive"]
+derive = ["multihash-derive"]
+arb = ["quickcheck", "rand"]
+secure-hashes = ["blake2b", "blake2s", "blake3", "sha2", "sha3"]
+scale-codec = ["parity-scale-codec"]
+serde-codec = ["serde", "generic-array/serde"]
+
+blake2b = ["blake2b_simd"]
+blake2s = ["blake2s_simd"]
+identity = []
+sha1 = ["digest", "sha-1"]
+sha2 = ["digest", "sha-2"]
+sha3 = ["digest", "sha-3"]
+strobe = ["strobe-rs"]
+
+[dependencies]
+generic-array = "0.14.4"
+parity-scale-codec = { version = "2.1.1", default-features = false, features = ["derive"], optional = true }
+quickcheck = { version = "0.9.2", optional = true }
+rand = { version = "0.7.3", optional = true }
+serde = { version = "1.0.116", default-features = false, features = ["derive"], optional = true }
+multihash-derive = { version = "0.1.0", path = "derive", optional = true }
+unsigned-varint = { version = "0.7.0", default-features = false }
+
+blake2b_simd = { version = "0.5.10", default-features = false, optional = true }
+blake2s_simd = { version = "0.5.10", default-features = false, optional = true }
+blake3 = { version = "0.3.7", default-features = false, optional = true }
+digest = { version = "0.9.0", default-features = false, optional = true }
+sha-1 = { version = "0.9.1", default-features = false, optional = true }
+sha-2 = { version = "0.9.0", default-features = false, optional = true, package = "sha2" }
+sha-3 = { version = "0.9.0", default-features = false, optional = true, package = "sha3" }
+strobe-rs = { version = "0.6.2", default-features = false, optional = true }
+
+sp-std = {version = "3", default-features = false}
+bytecursor = {path = "../bytecursor"}
+
+[dev-dependencies]
+criterion = "0.3.3"
+hex = "0.4.2"
+serde_json = "1.0.58"
+
+[[bench]]
+name = "multihash"
+harness = false
