+++ multihash/benches/multihash.rs
