+++ multihash/src/error.rs
