+++ multihash/src/hasher.rs
