+++ multihash/src/hasher_impl.rs
