--- sp-ipld/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "sp-ipld"
-version = "0.1.0"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-sp-std = { version = "3", default-features = false }
-byteorder = { version = "1", default-features = false }
-unsigned-varint = { version = "0.7.0", default-features = false }
-multibase = { version = "0.9.1", optional = true }
-parity-scale-codec = { version = "2.1.1", default-features = false, features = ["derive"], optional = true }
-serde = { version = "1.0.116", optional = true }
-
-bytecursor = { path = "../bytecursor" }
-sp-cid =  { path = "../sp-cid" }
-sp-multihash = { path = "../sp-multihash" }
-
-[dev-dependencies]
-quickcheck = "1.0.3"
-rand = "0.8.3"
-quickcheck_macros = "1.0.0"
-tokio = { version = "1.5", features = ["rt", "macros", "rt-multi-thread"] }
-reqwest = { version = "0.11.3", features = [ "multipart", "json" ] }
-serde_json = { version = "1.0" }
