--- sp-multihash/derive/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "sp-multihash-derive"
-version = "0.1.0"
-edition = "2018"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-proc-macro2 = { version = "1.0.24", features = ["span-locations"] }
-proc-macro-crate = "1.0.0"
-proc-macro-error = "1.0.4"
-quote = "1.0.7"
-syn = "1.0.42"
-synstructure = "0.12.4"
-
-[features]
-default = []
-std = []
-
-[dev-dependencies]
-pretty_assertions = "0.7.1"
-sp-multihash = { path = ".." }
